complwite

+++ etu/include/AC_DC.h
@@ -14,9 +14,9 @@ int16_t retourne_bits(int16_t nombre, uint8_t magnitude);
 
 
 void ecriture_symbole_DC(struct bitstream *stream, int16_t nombre, uint8_t indice);
-void ecriture_DC_y(struct bitstream *stream, struct mcu **tab, uint8_t indice_i, uint8_t indice_j ,uint8_t indice_k, int16_t predicateur);
-void ecriture_DC_cb(struct bitstream *stream, struct mcu **tab, uint8_t indice_i, uint8_t indice_j ,uint8_t indice_k, int16_t predicateur);
-void ecriture_DC_cr(struct bitstream *stream, struct mcu **tab, uint8_t indice_i, uint8_t indice_j ,uint8_t indice_k, int16_t predicateur);
+void ecriture_DC_y(struct bitstream *stream, struct mcu **tab, uint32_t indice_i, uint32_t indice_j ,uint8_t indice_k, int16_t predicateur);
+void ecriture_DC_cb(struct bitstream *stream, struct mcu **tab, uint32_t indice_i, uint32_t indice_j ,uint8_t indice_k, int16_t predicateur);
+void ecriture_DC_cr(struct bitstream *stream, struct mcu **tab, uint32_t indice_i, uint32_t indice_j ,uint8_t indice_k, int16_t predicateur);
 
 void AC_composante_puis_huffman(struct bitstream *stream, int16_t *composante, uint8_t id_y_cb_cr);
